Extract Crossref Metadata

Extracts Crossref metadata from the Crossref API using the item's Digital Object Identifier (DOI).

lib_metadata_crossref(doi, par = TRUE, ncores = NULL)

Arguments

doi
Character vector A vector of Digital Object Identifiers.

par
Logical. If TRUE, use multiple cores.

ncores
Integer. Number of cores to use if par = TRUE. If unspecified, defaults to detectCores() - 1.

Examples

# Single DOI
doi <- "https://doi.org/10.1016/j.adolescence.2014.04.012"
lib_metadata_crossref(
  doi = doi,
  par = FALSE
)

#>      doi                                
#> [1,] "10.1016/j.adolescence.2014.04.012"
#>      author                                                                
#> [1,] "Pesigan, Ivan Jacob Agaloos and Luyckx, Koen and Alampay, Liane Peña"
#>      editor type             
#> [1,] NA     "journal-article"
#>      title                                                                                                                           
#> [1,] "Brief report: Identity processes in Filipino late adolescents and young adults: Parental influences and mental health outcomes"
#>      original-title subtitle short-title publisher     publisher-location
#> [1,] NA             NA       NA          "Elsevier BV" NA                
#>      container-title          short-container-title    issue page      volume
#> [1,] "Journal of Adolescence" "Journal of Adolescence" "5"   "599-604" "37"  
#>      language ISNN
#> [1,] "en"     NA  
#>      subject                                                                                                                                         
#> [1,] "Pediatrics, Perinatology, and Child Health AND Developmental and Educational Psychology AND Psychiatry and Mental health AND Social Psychology"
#>      ISBN
#> [1,] NA  

doi <- "10.1016/j.adolescence.2014.04.012"
lib_metadata_crossref(
  doi = doi,
  par = FALSE
)

#>      doi                                
#> [1,] "10.1016/j.adolescence.2014.04.012"
#>      author                                                                
#> [1,] "Pesigan, Ivan Jacob Agaloos and Luyckx, Koen and Alampay, Liane Peña"
#>      editor type             
#> [1,] NA     "journal-article"
#>      title                                                                                                                           
#> [1,] "Brief report: Identity processes in Filipino late adolescents and young adults: Parental influences and mental health outcomes"
#>      original-title subtitle short-title publisher     publisher-location
#> [1,] NA             NA       NA          "Elsevier BV" NA                
#>      container-title          short-container-title    issue page      volume
#> [1,] "Journal of Adolescence" "Journal of Adolescence" "5"   "599-604" "37"  
#>      language ISNN
#> [1,] "en"     NA  
#>      subject                                                                                                                                         
#> [1,] "Pediatrics, Perinatology, and Child Health AND Developmental and Educational Psychology AND Psychiatry and Mental health AND Social Psychology"
#>      ISBN
#> [1,] NA  

# Vector of DOIs
doi <- c(
  "https://doi.org/10.1016/j.adolescence.2014.04.012",
  "https://doi.org/10.1007/s11469-015-9612-8"
)
lib_metadata_crossref(
  doi = doi,
  par = FALSE
)

#>      doi                                
#> [1,] "10.1016/j.adolescence.2014.04.012"
#> [2,] "10.1007/s11469-015-9612-8"        
#>      author                                                                
#> [1,] "Pesigan, Ivan Jacob Agaloos and Luyckx, Koen and Alampay, Liane Peña"
#> [2,] "Yu, Shu and Wu, Anise Man Sze and Pesigan, Ivan Jacob Agaloos"       
#>      editor type             
#> [1,] NA     "journal-article"
#> [2,] NA     "journal-article"
#>      title                                                                                                                           
#> [1,] "Brief report: Identity processes in Filipino late adolescents and young adults: Parental influences and mental health outcomes"
#> [2,] "Cognitive and Psychosocial Health Risk Factors of Social Networking Addiction"                                                 
#>      original-title subtitle short-title
#> [1,] NA             NA       NA         
#> [2,] NA             NA       NA         
#>      publisher                                 publisher-location
#> [1,] "Elsevier BV"                             NA                
#> [2,] "Springer Science and Business Media LLC" NA                
#>      container-title                                       
#> [1,] "Journal of Adolescence"                              
#> [2,] "International Journal of Mental Health and Addiction"
#>      short-container-title         issue page      volume language ISNN
#> [1,] "Journal of Adolescence"      "5"   "599-604" "37"   "en"     NA  
#> [2,] "Int J Ment Health Addiction" "4"   "550-564" "14"   "en"     NA  
#>      subject                                                                                                                                         
#> [1,] "Pediatrics, Perinatology, and Child Health AND Developmental and Educational Psychology AND Psychiatry and Mental health AND Social Psychology"
#> [2,] NA                                                                                                                                              
#>      ISBN
#> [1,] NA  
#> [2,] NA  

doi <- c(
  "10.1016/j.adolescence.2014.04.012",
  "10.1007/s11469-015-9612-8"
)
lib_metadata_crossref(
  doi = doi,
  par = FALSE
)

#>      doi                                
#> [1,] "10.1016/j.adolescence.2014.04.012"
#> [2,] "10.1007/s11469-015-9612-8"        
#>      author                                                                
#> [1,] "Pesigan, Ivan Jacob Agaloos and Luyckx, Koen and Alampay, Liane Peña"
#> [2,] "Yu, Shu and Wu, Anise Man Sze and Pesigan, Ivan Jacob Agaloos"       
#>      editor type             
#> [1,] NA     "journal-article"
#> [2,] NA     "journal-article"
#>      title                                                                                                                           
#> [1,] "Brief report: Identity processes in Filipino late adolescents and young adults: Parental influences and mental health outcomes"
#> [2,] "Cognitive and Psychosocial Health Risk Factors of Social Networking Addiction"                                                 
#>      original-title subtitle short-title
#> [1,] NA             NA       NA         
#> [2,] NA             NA       NA         
#>      publisher                                 publisher-location
#> [1,] "Elsevier BV"                             NA                
#> [2,] "Springer Science and Business Media LLC" NA                
#>      container-title                                       
#> [1,] "Journal of Adolescence"                              
#> [2,] "International Journal of Mental Health and Addiction"
#>      short-container-title         issue page      volume language ISNN
#> [1,] "Journal of Adolescence"      "5"   "599-604" "37"   "en"     NA  
#> [2,] "Int J Ment Health Addiction" "4"   "550-564" "14"   "en"     NA  
#>      subject                                                                                                                                         
#> [1,] "Pediatrics, Perinatology, and Child Health AND Developmental and Educational Psychology AND Psychiatry and Mental health AND Social Psychology"
#> [2,] NA                                                                                                                                              
#>      ISBN
#> [1,] NA  
#> [2,] NA