Download Resources from link.springer.com.

Downloads pdf and epub files from link.springer.com using wget. It is assumed that you have access to the resources via your library. It also assumes that wget is installed in your system.

lib_springer_download(
  dir,
  doi = NULL,
  chkfiles = TRUE,
  args = "-nc",
  remove_files = TRUE,
  type = c("pdf", "epub", "both"),
  par = TRUE,
  ncores = NULL
)

Arguments

dir
Character string. Directory which contains target files.

doi
Character vector. A vector of Digital Object Identifiers.

chkfiles
Logical. Check available files in dir. If TRUE, the download queue will EXCLUDE dois from available files.

args
Character string. Arguments to pass to wget.

remove_files
Logical. Remove files that do not match the specified file type. BE CAREFUL WITH THIS OPTION AS FILES CAN BE DELETED FROM YOUR SYSTEM.

type
Character string. Type of ebooks to download.

par
Logical. If TRUE, use multiple cores.

ncores
Integer. Number of cores to use if par = TRUE. If unspecified, defaults to detectCores() - 1.

Examples

# Single DOI
doi <- "10.1007/978-0-387-98141-3"
lib_springer_download(
  dir = getwd(),
  doi = doi,
  chkfiles = TRUE,
  type = "both",
  par = FALSE
)

#> Total number of unique DOIs provided: 1 
#> Checking downloaded files in /home/jek/scripts/r/jeksterslabRlib/docs/reference.
#> Downloading 1 PDF document/s...
#> Checking files and deleting non PDF documents...
#> Downloading 1 EPUB document/s...
#> Checking files and deleting non EPUB documents...
#> 1 available PDF documents.
#> 0 available EPUB documents.
#> See /home/jek/scripts/r/jeksterslabRlib/docs/reference/springer_books_available.Rds for available Springer books.

# Vector of DOIs
doi <- c(
  "https://doi.org/10.1007/978-0-387-71762-3",
  "10.1007/978-0-387-98141-3"
)
lib_springer_download(
  dir = getwd(),
  doi = doi,
  chkfiles = TRUE,
  type = "both",
  par = FALSE
)

#> Total number of unique DOIs provided: 2 
#> Checking downloaded files in /home/jek/scripts/r/jeksterslabRlib/docs/reference.
#> Downloading 1 PDF document/s...
#> Checking files and deleting non PDF documents...
#> Downloading 2 EPUB document/s...
#> Checking files and deleting non EPUB documents...
#> 2 available PDF documents.
#> 0 available EPUB documents.
#> See /home/jek/scripts/r/jeksterslabRlib/docs/reference/springer_books_available.Rds for available Springer books.